requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bert-score==0.3.13
bitsandbytes==0.46.0
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
DataProperty==1.1.0
datasets==3.6.0
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.8.1
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
latex2sympy2_extended==1.10.1
lxml==5.4.0
MarkupSafe==3.0.2
math-verify==0.7.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mistune==3.1.3
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
openai==1.76.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathvalidate==3.2.3
peft==0.15.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.7
portalocker==3.1.1
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyfiglet==1.0.3
Pygments==2.19.1
pyparsing==3.2.3
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
rpds-py==0.25.1
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.27.0
setproctitle==1.3.5
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
sqlitedict==2.1.0
stack-data==0.6.3
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
threadpoolctl==3.6.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
tornado==6.5.1
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.51.3
typepy==1.3.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.10
wcwidth==0.2.13
webencodings==0.5.1
word2number==1.1
xxhash==3.5.0
yarg==0.1.9
yarl==1.20.0
zstandard==0.23.0
sentencepiece==0.2.0
#####
psutil>=5.9.0 
nvidia-ml-py 
gputil>=1.4.0
seaborn
